+++ lib/rspec/sleeping_king_studios/concerns.rb
@@ -1,8 +1,11 @@
-# lib/rspec/sleeping_king_studios/concerns.rb
+# frozen_string_literal: true
 
 require 'rspec/sleeping_king_studios'
 
 module RSpec::SleepingKingStudios
   # RSpec-related concerns to mix into classes or objects.
-  module Concerns; end
-end # module
+  module Concerns
+    autoload :MemoizedHelpers,
+      'rspec/sleeping_king_studios/concerns/memoized_helpers'
+  end
+end